digital_twin/timescale/simpleRest/fastAPI_AQUC/requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
databases==0.5.5
energyusage==0.0.14
fastapi==0.75.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jsons==1.6.1
Pillow==9.1.0
psycopg2==2.9.3
pydantic==1.9.0
pyJoules==0.5.1
python-dotenv==0.19.2
reportlab==3.6.9
requests==2.27.1
sniffio==1.2.0
sockets==1.0.0
SQLAlchemy==1.4.32
starlette==0.17.1
thread6==0.2.0
typing_extensions==4.1.1
typish==1.9.3
urllib3==1.26.9
uvicorn==0.17.6
